feat: add params to toggle suppression of Unknown status scenarioruns

+++ scenariorun/src/main/kotlin/com/cosmotech/scenariorun/azure/ScenarioRunServiceImpl.kt
@@ -50,8 +50,6 @@ import com.cosmotech.solution.domain.Solution
 import com.cosmotech.workspace.api.WorkspaceApiService
 import com.cosmotech.workspace.domain.Workspace
 import com.fasterxml.jackson.databind.JsonNode
-import java.time.ZonedDateTime
-import kotlin.reflect.full.memberProperties
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
@@ -61,6 +59,8 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.event.EventListener
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Service
+import java.time.ZonedDateTime
+import kotlin.reflect.full.memberProperties
 
 private const val MIN_SDK_VERSION_MAJOR = 8
 private const val MIN_SDK_VERSION_MINOR = 5
@@ -142,49 +142,64 @@ internal class ScenarioRunServiceImpl(
     }
   }
 
-  override fun deleteHistoricalScenarioRunsByScenario(
+  override fun deleteHistoricalDataOrganization(organizationId: String, deleteUnknown: Boolean?) {
+    this.eventPublisher.publishEvent(
+        DeleteHistoricalDataOrganization(this, organizationId = organizationId, deleteUnknown))
+  }
+
+  override fun deleteHistoricalDataWorkspace(
       organizationId: String,
       workspaceId: String,
-      scenarioId: String
+      deleteUnknown: Boolean?
   ) {
+    this.eventPublisher.publishEvent(
+        DeleteHistoricalDataWorkspace(
+            this, organizationId = organizationId, workspaceId = workspaceId, deleteUnknown))
+  }
 
+  override fun deleteHistoricalDataScenario(
+      organizationId: String,
+      workspaceId: String,
+      scenarioId: String,
+      deleteUnknown: Boolean?
+  ) {
     GlobalScope.launch {
       this@ScenarioRunServiceImpl.deleteScenarioRunsByScenarioWithoutAccessEnforcement(
-          organizationId, workspaceId, scenarioId)
+          organizationId, workspaceId, scenarioId, deleteUnknown)
     }
   }
 
   private fun deleteScenarioRunsByScenarioWithoutAccessEnforcement(
       organizationId: String,
       workspaceId: String,
-      scenarioId: String
+      scenarioId: String,
+      deleteUnknown: Boolean?
   ) {
     val scenarioRuns = getScenarioRuns(organizationId, workspaceId, scenarioId).toMutableList()
 
+      val lastRunId =
+          scenarioApiService.findScenarioById(organizationId, workspaceId, scenarioId).lastRun!!
+              .scenarioRunId
+
     scenarioRuns.filter { it.state == ScenarioRunState.Failed }.forEach {
-      deleteScenarioRunWithoutAccessEnforcement(it)
+        if (it.id != lastRunId) {
+            deleteScenarioRunWithoutAccessEnforcement(it)
+        }
     }
 
-    val lastRunId =
-        scenarioApiService.findScenarioById(organizationId, workspaceId, scenarioId).lastRun!!
-            .scenarioRunId
-
-    for (run in scenarioRuns.filter { it.state == ScenarioRunState.Successful }) {
-      if (run.id != lastRunId) {
-        deleteScenarioRunWithoutAccessEnforcement(run)
+      if(deleteUnknown!!) {
+          scenarioRuns.filter { it.state == ScenarioRunState.Unknown }.forEach {
+              if (it.id != lastRunId) {
+                  deleteScenarioRunWithoutAccessEnforcement(it)
+              }
+          }
       }
+
+    scenarioRuns.filter { it.state == ScenarioRunState.Successful }.forEach {
+        if (it.id != lastRunId) {
+            deleteScenarioRunWithoutAccessEnforcement(it)
+        }
     }
-  }
-
-  override fun deleteHistoricalDataOrganization(organizationId: String) {
-    this.eventPublisher.publishEvent(
-        DeleteHistoricalDataOrganization(this, organizationId = organizationId))
-  }
-
-  override fun deleteHistoricalDataWorkspace(organizationId: String, workspaceId: String) {
-    this.eventPublisher.publishEvent(
-        DeleteHistoricalDataWorkspace(
-            this, organizationId = organizationId, workspaceId = workspaceId))
   }
 
   override fun findScenarioRunById(organizationId: String, scenariorunId: String) =
@@ -309,7 +324,7 @@ internal class ScenarioRunServiceImpl(
 
   @EventListener(DeleteHistoricalDataScenario::class)
   fun deleteHistoricalDataScenarioRun(data: DeleteHistoricalDataScenario) {
-    deleteHistoricalScenarioRunsByScenario(data.organizationId, data.workspaceId, data.scenarioId)
+    deleteHistoricalDataScenario(data.organizationId, data.workspaceId, data.scenarioId, false)
   }
 
   @EventListener(ScenarioDataDownloadRequest::class)
@@ -426,7 +441,7 @@ internal class ScenarioRunServiceImpl(
     }
     if (scenarioRunStatus == ScenarioRunState.Successful.value) {
       logger.info("ScenarioRun {} is Successfull => purging data", csmSimulationRun)
-      deleteScenarioRunsByScenarioWithoutAccessEnforcement(organizationId, workspaceId, scenarioId)
+      deleteScenarioRunsByScenarioWithoutAccessEnforcement(organizationId, workspaceId, scenarioId, false)
     } else {
       logger.info(
           "ScenarioRun {} is in error {} => no purging data", csmSimulationRun, scenarioRunStatus)
